% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_sequence.R
\name{date.sequence}
\alias{date.sequence}
\title{SST data extracted to spatial points}
\usage{
date.sequence(year, month, day, day.buffer)
}
\arguments{
\item{year}{A character string.}

\item{month}{An integer.}

\item{day}{An integer.}

\item{buffer}{An integer. The number of days on each side of the date you want
returned as a list of dates.}
}
\value{
A sequence of dates
}
\description{
SST data extracted to spatial points
}
\examples{
date_seq <- date.sequence(year="2015", month="01", day="10", day.buffer=30)

}
