% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sst.R
\name{extract_sst}
\alias{extract_sst}
\title{Extract SST for a specific date/time}
\usage{
extract_sst(year = "2013", month = "01", day = "01", bbox = c(-15, 10, 45, 62))
}
\arguments{
\item{year}{A character string.}

\item{month}{An integer.}

\item{day}{An integer.}

\item{bbox}{(optional) A vector in the format Xmin, Xmax, Ymin, Ymax for plotting.}

\item{plot.it}{A boolean. TRUE will plot the output with a contoured levelplot}
}
\value{
A raster object
}
\description{
Extract SST for a specific date/time
}
\examples{
r <- extract_sst(year,month,day,bbox=c(-15,10,45,62))

}
