% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_to_raster.R
\name{nc_to_raster}
\alias{nc_to_raster}
\title{Reads a netCDF file and returns a raster}
\usage{
nc_to_raster(filepath, bbox = c(-15, 10, 45, 62))
}
\arguments{
\item{filepath}{A character string for the filename}

\item{bbox}{(optional) A vector in the format Xmin, Xmax, Ymin, Ymax for plotting.}
}
\value{
A raster object
}
\description{
Reads a netCDF file and returns a raster
}
\examples{
r <- nc_to_raster(filepath="/ncdata.ncf",bbox=c(-15,10,45,62))

}
