% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsst.R
\name{plot.sst}
\alias{plot.sst}
\title{Plot one NCDF file downloaded from GHRSST}
\usage{
plot.sst(filepath, bbox = c(-15, 10, 45, 62))
}
\arguments{
\item{filepath}{A character string for the filename}

\item{bbox}{(optional) A vector in the format Xmin, Xmax, Ymin, Ymax for plotting.}
}
\value{
A levelplot object from the rasterVis library
}
\description{
Reads a netCDF file and the plots the output
}
\examples{
P <- plot.sst(file="/ncdata.ncf",bbox=c(-15,10,45,62))
}
