% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_aggregation.R
\name{raster_aggregation}
\alias{raster_aggregation}
\title{Create mean, median and sd rasters of SST from a sequence of dates}
\usage{
raster_aggregation(DateS)
}
\arguments{
\item{Dates}{A sequence of dates output from date.sequence()}
}
\value{
A raster stack of 3 raster layers (mean, median and standard deviation)
}
\description{
Create mean, median and sd rasters of SST from a sequence of dates
}
\examples{

year = "2018"
month = "06"
day = "10"
DateS <- date.sequence(year,month,day,7)
Rs <- raster_aggregation(DateS)

}
